+++ src/exception.cpp
@@ -13,7 +13,7 @@ void ExceptionHandler::EH_abort(){
     abort();
 }
 
-void ExceptionHandler::CheckForArithmeticOverflow(const int32_t operand1, const int32_t operand2, const int32_t opcode) {
+void ExceptionHandler::CheckForArithmeticOverflow(const int64_t operand1, const int64_t operand2, const int32_t opcode) {
     if((operand2 > 0 && operand1 > (INT32_MAX - operand2)) || (operand2 < 0 && operand1 < (INT32_MAX - operand2)))
         Handler(EXCEPTION_ARITHMETIC_OVERFLOW, opcode);
 }
